requirements.txt
_libgcc_mutex             0.1                     
_openmp_mutex             4.5                      
absl-py                   0.15.0            
aiohttp                   3.7.4.post0       
ase                       3.22.0            
async-timeout             3.0.1             
attrs                     21.2.0            
ca-certificates           2021.9.30          
cached-property           1.5.2             
cachetools                4.2.4             
certifi                   2021.10.8        
chardet                   4.0.0             
charset-normalizer        2.0.7             
cycler                    0.10.0            
cython                    0.29.24           
fsspec                    2021.10.1         
future                    0.18.2            
google-auth               2.3.2             
google-auth-oauthlib      0.4.6             
googledrivedownloader     0.4               
grpcio                    1.41.1            
h5py                      3.5.0             
idna                      3.3               
importlib-metadata        4.8.1             
isodate                   0.6.0             
jinja2                    3.0.2             
joblib                    1.1.0             
kiwisolver                1.3.2             
ld_impl_linux-64          2.35.1              
libffi                    3.3                  
libgcc-ng                 9.3.0              
libgomp                   9.3.0              
libstdcxx-ng              9.3.0              
littleutils               0.2.2             
llvmlite                  0.37.0            
markdown                  3.3.4             
markupsafe                2.0.1             
matplotlib                3.4.3             
multidict                 5.2.0             
ncurses                   6.2                
networkx                  2.6.3             
numba                     0.54.1            
numpy                     1.20.3            
oauthlib                  3.1.1             
ogb                       1.3.1             
opencv-python             4.6.0.66          
openssl                   1.1.1l               
outdated                  0.2.1             
packaging                 21.0              
pandas                    1.3.4             
pillow                    8.4.0             
pip                       21.0.1          
protobuf                  3.19.0            
pyasn1                    0.4.8             
pyasn1-modules            0.2.8             
pydeprecate               0.3.0             
pyparsing                 3.0.1             
python                    3.7.11              
python-dateutil           2.8.2             
python-louvain            0.15              
pytorch-lightning         1.3.1             
pytz                      2021.3            
pyyaml                    5.4.1             
rdflib                    6.0.2             
rdkit-pypi                2021.9.1          
readline                  8.1               
requests                  2.26.0            
requests-oauthlib         1.3.0             
rsa                       4.7.2             
scikit-learn              1.0.1             
scipy                     1.7.1             
setuptools                58.0.4           
six                       1.16.0            
sqlite                    3.36.0               
tensorboard               2.7.0             
tensorboard-data-server   0.6.1             
tensorboard-plugin-wit    1.8.0             
tensorboardx              2.5.1             
threadpoolctl             3.0.0             
tk                        8.6.11              
torch                     1.11.0+cu113      
torch-geometric           1.6.3             
torch-scatter             2.0.6             
torch-sparse              0.6.9             
torchmetrics              0.5.1             
torchvision               0.12.0+cu113      
tqdm                      4.62.3            
typing-extensions         3.10.0.2          
urllib3                   1.26.7            
werkzeug                  2.0.2             
wheel                     0.37.0            
xz                        5.2.5            
yarl                      1.7.0             
zipp                      3.6.0             
zlib                      1.2.11              